Module Ppx_repr_lib

module Engine : sig ... end

Derivers and expanders for ppx_repr.

module Plugins : sig ... end

Functions for defining PPXes that use the implementation defined in Engine.

module Meta_deriving : sig ... end

A meta-deriver is a deriver defined in terms of specialisations of Repr's generic operations (i.e. with plugins that consume type representations rather than type ASTs).