Module Algebraic_intf.Typ

type nonrec record_field_repr = {

field_name : string;

field_repr : Ppxlib.expression;

}

and variant_case_repr = {

case_name : string;

case_cons : (Ppxlib.expression * int) option;

}

type (_, _) t =

| Record : (Ppxlib.label_declaration, record_field_repr) t

| Variant : (Ppxlib.constructor_declaration, variant_case_repr) t

| Polyvariant : (Ppxlib.row_field, variant_case_repr) t

The algebraic datatypes supported by this module, parameterised by:

'a: the subcomponent type of the algebraic type

'b: a generic representation of the subcomponent type necessary to derive the composite type representation