Module Ppx_repr_lib.Dsl

This module defines the expected combinators exposed by Repr, and their correspondences to the types in the standard library.

val drop_stdlib_prefix : Ppxlib.longident -> Ppxlib.longident

val basic : string list

val containers : string list

val type_in_default_scope : string -> Ppxlib.longident * string

val type_in_separate_module : string -> Ppxlib.longident * string

val type_to_combinator_name : Ppxlib.longident -> string option